requirements.txt
Django~=5.0.4
pip~=22.3.1
pytest~=8.1.1
asgiref~=3.8.1
sqlparse~=0.5.0
pluggy~=1.5.0
iniconfig~=2.0.0
setuptools~=65.5.0
packaging~=24.0
djangorestframework~=3.15.1